ref(metrics): Move metrics related modules into proper metrics module
