+++ relay-server/src/metrics/mod.rs
@@ -1,0 +1,13 @@
+#[cfg(feature = "processing")]
+mod bucket_encoding;
+mod metric_stats;
+mod minimal;
+mod outcomes;
+mod rate_limits;
+
+#[cfg(feature = "processing")]
+pub use self::bucket_encoding::*;
+pub use self::metric_stats::*;
+pub use self::minimal::*;
+pub use self::outcomes::*;
+pub use self::rate_limits::*;
