+++ relay-server/src/metrics/outcomes.rs
@@ -7,15 +7,10 @@ use relay_quotas::{DataCategory, Scoping};
 use relay_system::Addr;
 
 use crate::envelope::SourceQuantities;
-use crate::metric_stats::MetricStats;
+use crate::metrics::{ExtractionMode, MetricStats};
 use crate::services::outcome::{Outcome, TrackOutcome};
-use crate::utils::ExtractionMode;
 #[cfg(feature = "processing")]
 use relay_cardinality::{CardinalityLimit, CardinalityReport};
-
-mod minimal;
-
-pub use self::minimal::*;
 
 pub const PROFILE_TAG: &str = "has_profile";
