+++ relay-server/src/utils/mod.rs
@@ -1,11 +1,8 @@
 mod api;
-#[cfg(feature = "processing")]
-mod bucket_encoding;
 mod buffer;
 mod dynamic_sampling;
 mod garbage;
 mod managed_envelope;
-mod metrics_rate_limits;
 mod multipart;
 mod param_parser;
 mod pick;
@@ -23,13 +20,10 @@ mod native;
 mod unreal;
 
 pub use self::api::*;
-#[cfg(feature = "processing")]
-pub use self::bucket_encoding::*;
 pub use self::buffer::*;
 pub use self::dynamic_sampling::*;
 pub use self::garbage::*;
 pub use self::managed_envelope::*;
-pub use self::metrics_rate_limits::*;
 pub use self::multipart::*;
 #[cfg(feature = "processing")]
 pub use self::native::*;
